requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
pytest>=7.4.0
loguru>=0.7.0
tqdm>=4.66.0
scikit-learn>=1.3.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
streamlit>=1.30.0
plotly>=5.18.0
